build: move herb derivation to nix/pkgs
A useful convention is for package sources to live in 'pkg' and package
derivations to live in 'nix/pkgs'.  Subdirectories of 'nix/pkgs' are
then a collection of things that we might want to release.

herb deviated from this norm, but now follows the same pattern as e.g.
arvo and vere.